Rename as topic-parse for consistent naming with broadcast messages middleware
